+++ app-web/src/components/Cards/Card/Card.js
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faExternalLinkSquareAlt } from '@fortawesome/free-solid-svg-icons';
 import styles from './Card.module.css';
 import Link from '../../UI/Link/Link';
-import { ARIA_LABEL_RESOURCE, ARIA_LABEL_REPO } from '../../../constants/strings';
+import { ARIA_LABEL_RESOURCE, ARIA_LABEL_REPO } from '../../../constants/ariaLabels';
 
 const Card = ({ sourcePath, sourceName, title, description, resourcePath }) => (
   <article className={styles.Card}>
