+++ app-web/src/components/PrimaryHeader/PrimaryHeader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Flag } from 'flag';
 import classes from './PrimaryHeader.module.css';
-import { APP_TITLE } from '../../constants/strings';
+import { APP_TITLE } from '../../constants/ariaLabels';
 import Banner from '../UI/Banner/Banner';
 import Login from '../Auth/Login/Login';
 import PrimaryFilter from '../PrimaryFilter/PrimaryFilter';
