+++ app-web/src/components/UI/Hamburger/Hamburger.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
 import PropTypes from 'prop-types';
-import { ARIA_LABEL_FILTER_SELECT } from '../../../constants/strings';
+import { ARIA_LABEL_FILTER_SELECT } from '../../../constants/ariaLabels';
 const Hamburger = ({ clicked, ...rest }) => (
   <FontAwesomeIcon
     {...rest}
